rename `table_state` folder to `state_cache`
