+++ src/stream/src/executor/aggregation/state_cache/array_agg.rs
