+++ src/stream/src/executor/aggregation/state_cache/extreme.rs
