+++ src/stream/src/executor/aggregation/state_cache/string_agg.rs
